# Bug 1689167: Reduce memory usage in probe scraper