Probe scraper loads all probes over all time into memory, before slicing the information in a few different ways and writing it back out.  I measured it using 30GB of RAM before I gave up on it because my machine slowed to a crawl.

There are probably some more clever ways to avoid loading all the data at once, instead streaming data in and out and making greater use of the disk.

However, as a stop gap measure, the approach in this PR seems to work quite well.  Even though there are many (3,400) revisions of the probe metadata files, the individual probes for the most part don't change at every revision.  Therefore, we can save a lot of memory by reusing a probe when we've already seen one with the exact same contents.

To test the impact of this, I ran this over the nightly channel only (running all channels would have taken much more testing time, and nightly represents a "worst case").  With this optimization, the maximum memory use is 1.2 GB vs. 17.5 GB.  Runtime is also slightly lower: 22 min vs. 28 min wall clock time.  Obviously Airflow will behave differently than my local machine, but I think this a solid improvement for something that's pretty localized.

Note: the Glean metrics parsing code path is completely separate.  It will have this same problem *some day* but for now the revisions and number of probes remains much lower.

# Before:

```
1638.12s user 18.49s system 98% cpu 27:57.03 total
```

![image](https://user-images.githubusercontent.com/38294/106191442-defca080-6178-11eb-87b5-01ea619436d2.png)

# After:

```
1300.64s user 3.92s system 99% cpu 21:54.72 total
```

![image](https://user-images.githubusercontent.com/38294/106191559-03587d00-6179-11eb-9529-eadf7649993f.png)